requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
Django==5.2.3
djangorestframework==3.16.0
dotenv==0.9.9
drf-yasg==1.21.10
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-telegram-bot==22.1
pytz==2025.2
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
